NEXT-12159 - Refactor Elasticsearch indexing

See merge request shopware/6/product/platform!4499

+++ src/Elasticsearch/Profiler/DataCollector.php
@@ -41,7 +41,7 @@ class DataCollector extends BaseDataCollector
         $this->client->resetRequests();
     }
 
-    public function getTime(): int
+    public function getTime(): float
     {
         $time = 0;
